requirements.txt
requests
flask
PySide6
python-dotenv
pyinstaller